Use uint32 in vanilla voting strategy also

+++ src/voting-strategies/VanillaVotingStrategy.sol
@@ -5,7 +5,7 @@ pragma solidity ^0.8.15;
 import "../interfaces/IVotingStrategy.sol";
 
 contract VanillaVotingStrategy is IVotingStrategy {
-    function getVotingPower(uint256, address, bytes memory, bytes memory) external pure returns (uint256 votingPower) {
+    function getVotingPower(uint32, address, bytes memory, bytes memory) external pure returns (uint256 votingPower) {
         return 1;
     }
 }
